cleanup, add migrations
